use global postMessage
